+++ adalflow/adalflow/optim/optimizer.py
@@ -1,12 +1,12 @@
 """Base Classes for AdalFlow Optimizers, including Optimizer, TextOptimizer, and DemoOptimizer."""
 
-from typing_extensions import TypeAlias
-from typing import Dict, Any, Union, Iterable, Sequence
+from typing import Dict, Any, Union, Iterable, Sequence, List
 
 from adalflow.optim.parameter import Parameter
 from adalflow.core.base_data_class import DataClass
 
-ParamsT: TypeAlias = Union[Iterable[Parameter], Iterable[Dict[str, Any]]]
+
+ParamsT = Union[Iterable[Parameter], Iterable[Dict[str, Any]]]
 
 
 class Optimizer:
@@ -85,3 +85,7 @@ class DemoOptimizer(Optimizer):
     def set_dataset(self, dataset: Sequence[DataClass]):
         r"""Set the dataset for the optimizer."""
         self.dataset = dataset
+
+    def add_scores(self, ids: List[str], scores: List[float], *args, **kwargs):
+        r"""Add scores to the optimizer."""
+        raise NotImplementedError("add_scores method is not implemented")