+++ adalflow/adalflow/optim/trainer/adal.py
@@ -474,9 +474,7 @@ class AdalComponent(Component):
         )
         if index_to_score:
             # compute score from index_to_score
-            print(
-                f"completed_samples: {len(completed_samples)}, len: {len(list(index_to_score.values()))}"
-            )
+
             avg_score = np.mean(list(index_to_score.values())).item()
             acc_list = [None] * len(index_to_score)
             for i, score in index_to_score.items():
@@ -598,7 +596,9 @@ class AdalComponent(Component):
         if self.loss_fn:
             self.loss_fn.set_backward_engine(self.backward_engine)
 
-    def configure_callbacks(self, save_dir: Optional[str] = "traces", *args, **kwargs):
+    def configure_callbacks(
+        self, save_dir: Optional[str] = "traces", *args, **kwargs
+    ) -> List[str]:
         """In default we config the failure generator callback. User can overwrite this method to add more callbacks."""
         from adalflow.utils.global_config import get_adalflow_default_root_path
         import os
@@ -606,7 +606,7 @@ class AdalComponent(Component):
         if not save_dir:
             save_dir = "traces"
             save_dir = os.path.join(get_adalflow_default_root_path(), save_dir)
-        print(f"Saving traces to {save_dir}")
+        log.debug(f"Saving traces to {save_dir}")
         return self._auto_generator_callbacks(save_dir)
 
     def run_one_task_sample(self, sample: Any) -> Any:
@@ -640,9 +640,10 @@ class AdalComponent(Component):
         for name, comp in self.task.named_components():
             if isinstance(comp, Generator):
                 all_generators.append((name, comp))
+        log.debug(f"all_generators: {all_generators}")
         return all_generators
 
-    def _auto_generator_callbacks(self, save_dir: str = "traces"):
+    def _auto_generator_callbacks(self, save_dir: str = "traces") -> List[str]:
         r"""Automatically generate callbacks."""
         from adalflow.core.types import GeneratorOutput
         from adalflow.tracing.generator_call_logger import (
@@ -652,7 +653,7 @@ class AdalComponent(Component):
 
         all_generators = self._find_all_generators()
 
-        print(f"all_generators: {all_generators}")
+        log.debug(f"all_generators: {all_generators}")
 
         def _on_completion_callback(
             output: GeneratorOutput,
@@ -672,9 +673,10 @@ class AdalComponent(Component):
         # Register the callback for each generator
 
         file_paths = []
+        call_logger = GeneratorCallLogger(save_dir=save_dir)
         for name, generator in all_generators:
-            call_logger = GeneratorCallLogger(save_dir=save_dir)
-            call_logger.reset()
+
+            # call_logger.reset()
             call_logger.register_generator(name)
             logger_call = partial(call_logger.log_call, name)
             generator.register_callback(
@@ -682,10 +684,7 @@ class AdalComponent(Component):
             )
             file_path = call_logger.get_log_location(name)
             file_paths.append(file_path)
-            print(
-                f"Registered callback for {name}, file path: {file_path}",
-                end="\n",
-            )
+            log.debug(f"Registered callback for {name}, file path: {file_path}")
         return file_paths
 
     def configure_demo_optimizer_helper(self) -> List[DemoOptimizer]: